[refactor] extract seat selection logic into SeatService for better separation of concerns
